refactor(dbaas): move role and acl into dedicated packages
Split the dbaas package: role and acl subcommand groups now live in
cmd/dbaas/role and cmd/dbaas/acl, following the subpackage layout used
by cmd/storage/lifecycle and cmd/kms/key. The parent dbaas package
attaches their Cmd in init, and their tests move along.

AI-assisted: true
