+++ cmd/dbaas/acl/acl.go
@@ -1,0 +1,12 @@
+// AI-generated by qwen3.8-27b-nvfp4-dflash2 - not reviewed yet
+package acl
+
+import (
+	"github.com/spf13/cobra"
+)
+
+// Cmd is the root command for DBaaS ACL subcommands.
+var Cmd = &cobra.Command{
+	Use:   "acl",
+	Short: "Manage DBaaS ACL configuration",
+}
