+++ cmd/dbaas/acl/acl_show.go
@@ -1,4 +1,5 @@
-package dbaas
+// AI-generated by qwen3.8-27b-nvfp4-dflash2 - not reviewed yet
+package acl
 
 import (
 	"bytes"
@@ -148,7 +149,7 @@ func (o *dbaasAclShowOutput) ToTable() {
 }
 
 func init() {
-	cobra.CheckErr(exocmd.RegisterCLICommand(dbaasAclCmd, &dbaasAclShowCmd{
+	cobra.CheckErr(exocmd.RegisterCLICommand(Cmd, &dbaasAclShowCmd{
 		CliCommandSettings: exocmd.DefaultCLICmdSettings(),
 	}))
 }
