--- cmd/dbaas/dbaas_acl.go
@@ -1,14 +1,0 @@
-package dbaas
-
-import (
-	"github.com/spf13/cobra"
-)
-
-var dbaasAclCmd = &cobra.Command{
-	Use:   "acl",
-	Short: "Manage DBaaS ACL configuration",
-}
-
-func init() {
-	dbaasCmd.AddCommand(dbaasAclCmd)
-}
