+++ cmd/dbaas/role/role.go
@@ -1,0 +1,12 @@
+// AI-generated by qwen3.8-27b-nvfp4-dflash2 - not reviewed yet
+package role
+
+import (
+	"github.com/spf13/cobra"
+)
+
+// Cmd is the root command for DBaaS role subcommands.
+var Cmd = &cobra.Command{
+	Use:   "role",
+	Short: "Manage DBaaS roles",
+}
