+++ cmd/dbaas/role/role_delete.go
@@ -1,4 +1,5 @@
-package dbaas
+// AI-generated by qwen3.8-27b-nvfp4-dflash2 - not reviewed yet
+package role
 
 import (
 	"fmt"
@@ -67,7 +68,7 @@ func (c *dbaasRoleDeleteCmd) CmdRun(_ *cobra.Command, _ []string) error {
 }
 
 func init() {
-	cobra.CheckErr(exocmd.RegisterCLICommand(dbaasRoleCmd, &dbaasRoleDeleteCmd{
+	cobra.CheckErr(exocmd.RegisterCLICommand(Cmd, &dbaasRoleDeleteCmd{
 		CliCommandSettings: exocmd.DefaultCLICmdSettings(),
 	}))
 }
