+++ cmd/dbaas/role/role_list.go
@@ -1,4 +1,5 @@
-package dbaas
+// AI-generated by qwen3.8-27b-nvfp4-dflash2 - not reviewed yet
+package role
 
 import (
 	"bytes"
@@ -153,7 +154,7 @@ func (o *dbaasRoleListOutput) ToTable() {
 }
 
 func init() {
-	cobra.CheckErr(exocmd.RegisterCLICommand(dbaasRoleCmd, &dbaasRoleListCmd{
+	cobra.CheckErr(exocmd.RegisterCLICommand(Cmd, &dbaasRoleListCmd{
 		CliCommandSettings: exocmd.DefaultCLICmdSettings(),
 	}))
 }
